2.0.0 - switched the whole structure tot the smart pointers

+++ src/generator/dna_generator.cpp
@@ -1,10 +1,8 @@
 //
 // Created by konrad_guest on 28/12/2024.
-//
+// SMART
 
 #include "generator/dna_generator.h"
-
-
 
 /* **********************************************
  *          RandomGenerator – singleton
@@ -40,8 +38,7 @@ std::string DNAGenerator::generateDNA(int n, bool repAllowed)
 
     // Prosty wariant – generujemy dowolne znaki ACGT.
     // Jeśli repAllowed = false, można np. kontrolować,
-    // by nie było 2-3 identycznych znaków pod rząd, itp. (wedle uznania).
-    // Poniżej najprostsze rozwiązanie bez rozbudowanych reguł:
+    // by nie było 2-3 identycznych znaków pod rząd, itp.
     for(int i = 0; i < n; ++i) {
         dna.push_back(nucleotides[dist(rng)]);
     }
@@ -68,30 +65,19 @@ std::vector<std::string> SpectrumGenerator::generateSpectrum(const std::string &
     int startPos = 0;
     int currentOligoLength = k;
 
-    // Ile w sumie powstanie oligonukleotydów = dna.size() - k + 1, ale z uwzględnieniem,
-    // że "okno" przesuwamy o 1 za każdym razem.
-    // Natomiast ostatnich (k+2) pozycji ma zawsze długość k.
-    // Przykładowo – jeśli DNA ma 100 bp i k=8, to maksymalny indeks startowy to 92.
-    // Ostatnie k+2 = 10 pozycji, więc od 90 do 99 włącznie.
-    // Dlatego do momentu ( dna.size() - (k+2) ) wylosujemy ewentualnie inne długości.
-
     int endFixedZoneStart = dna.size() - (k + 2);
     if(endFixedZoneStart < 0) {
-        // Gdyby k+2 > dna.size(), wówczas nie ma "zmiennej" części, wszystko jest stałe.
         endFixedZoneStart = 0;
     }
 
     while (startPos + k <= (int)dna.size()) {
         // Sprawdzamy, czy jesteśmy już w strefie ostatnich (k+2) oligonów
         if (startPos <= endFixedZoneStart) {
-            // Dla pierwszego (startPos == 0) zachowujemy k
             if (startPos == 0) {
                 currentOligoLength = k;
             } else {
-                // Losujemy wartość z zakresu [0..deltaK]
                 int d = distDelta(rng);
                 if (d > 0) {
-                    // Losujemy znak
                     int sign = distSign(rng);
                     if (sign == 0) {
                         currentOligoLength = std::max(k - d, 1); 
@@ -99,28 +85,22 @@ std::vector<std::string> SpectrumGenerator::generateSpectrum(const std::string &
                         currentOligoLength = std::min(k + d, (int)dna.size() - startPos);
                     }
                 } else {
-                    // D == 0 => nic nie zmieniamy
                     currentOligoLength = k;
                 }
             }
         } else {
-            // W strefie końcowej zawsze bierzemy oligo o długości k,
-            // o ile wystarczy znaków do wzięcia
             if (startPos + k > (int)dna.size()) {
                 break;
             }
             currentOligoLength = k;
         }
 
-        // Zabezpieczenie, żeby nie wyjść za koniec łańcucha
         if (startPos + currentOligoLength > (int)dna.size()) {
             break;
         }
 
-        // Wycinamy fragment
         spectrum.push_back(dna.substr(startPos, currentOligoLength));
 
-        // Przesuwamy okno o 1
         startPos += 1;
     }
 
@@ -137,7 +117,6 @@ bool InstanceIO::saveInstance(const DNAInstance &instance, const std::string &fi
         return false;
     }
 
-    // Zapisujemy: n, k, deltaK, lNeg, lPoz, repAllowed, probablePositive
     out << instance.getN() << " " 
         << instance.getK() << " "
         << instance.getDeltaK() << " "
@@ -146,10 +125,8 @@ bool InstanceIO::saveInstance(const DNAInstance &instance, const std::string &fi
         << (instance.isRepAllowed() ? 1 : 0) << " "
         << instance.getProbablePositive() << "\n";
 
-    // Zapisujemy DNA
     out << instance.getDNA() << "\n";
 
-    // Zapisujemy rozmiar spektrum i poszczególne k-mery
     out << instance.getSpectrum().size() << "\n";
     for (const auto &frag : instance.getSpectrum()) {
         out << frag << "\n";
@@ -200,9 +177,7 @@ bool InstanceIO::loadInstance(const std::string &filename, DNAInstance &instance
  * **********************************************/
 DNAInstanceBuilder& DNAInstanceBuilder::setN(int n)
 {
-    // Walidacja, np. min = 300, max = 700
     if(n < 300 || n > 700) {
-        // Można rzucić wyjątek, lub ustawić na domyślną wartość
         n = 400;
     }
     instance.setN(n);
@@ -211,7 +186,6 @@ DNAInstanceBuilder& DNAInstanceBuilder::setN(int n)
 
 DNAInstanceBuilder& DNAInstanceBuilder::setK(int k)
 {
-    // Walidacja, np. min = 7, max = 10
     if(k < 7 || k > 10) {
         k = 8;
     }
@@ -221,7 +195,6 @@ DNAInstanceBuilder& DNAInstanceBuilder::setK(int k)
 
 DNAInstanceBuilder& DNAInstanceBuilder::setDeltaK(int dk)
 {
-    // Walidacja, np. min = 0, max = 2
     if(dk < 0 || dk > 2) {
         dk = 2;
     }
@@ -231,24 +204,17 @@ DNAInstanceBuilder& DNAInstanceBuilder::setDeltaK(int dk)
 
 DNAInstanceBuilder& DNAInstanceBuilder::setLNeg(int ln)
 {
-    // l_neg może być 0 lub minimum 10,
-    // nie więcej niż ~15% wszystkich oligonukleotydów (to sprawdza się dopiero po wytworzeniu?)
-    // Na potrzeby przykładu – wstępna walidacja:
     if(ln != 0 && ln < 10) {
-        ln = 0;  // lub throw
+        ln = 0;
     }
     instance.setLNeg(ln);
     return *this;
 }
 
 DNAInstanceBuilder& DNAInstanceBuilder::setLPoz(int lp) {
-    // Sprawdzamy czy liczba błędów pozytywnych jest rozsądna
     if(lp != 0) {
-        // Minimum 10 błędów jeśli nie zero
         if(lp < 10) lp = 10;
-        
-        // Ograniczamy maksymalną liczbę błędów do ~50% obecnego spektrum
-        // aby uniknąć zbyt dużego rozrostu
+
         if(!instance.getSpectrum().empty()) {
             int maxErrors = instance.getSpectrum().size() / 2;
             if(lp > maxErrors) lp = maxErrors;
@@ -260,9 +226,8 @@ DNAInstanceBuilder& DNAInstanceBuilder::setLPoz(int lp) {
 
 DNAInstanceBuilder& DNAInstanceBuilder::setRepAllowed(bool rep)
 {
-    // Parametr ignorowany, jeśli l_neg > 0
     if(instance.getLNeg() > 0) {
-        rep = true;  // lub false – zależnie od interpretacji; w treści "nieistotne"
+        rep = true;
     }
     instance.setRepAllowed(rep);
     return *this;
@@ -270,7 +235,6 @@ DNAInstanceBuilder& DNAInstanceBuilder::setRepAllowed(bool rep)
 
 DNAInstanceBuilder& DNAInstanceBuilder::setProbablePositive(int val)
 {
-    // 0 lub 1
     if(val != 0 && val != 1) {
         val = 0;
     }
@@ -295,9 +259,6 @@ DNAInstanceBuilder& DNAInstanceBuilder::buildSpectrum()
         instance.getDeltaK()
     );
     instance.setSpectrum(spec);
-
-    // Tutaj można dodatkowo sprawdzić czy lNeg <= 15% wielkości spektrum, itp.
-    // lub ewentualnie skorygować te wartości.
 
     return *this;
 }
@@ -324,7 +285,6 @@ void NegativeErrorIntroducer::introduceErrors(DNAInstance &instance)
         return;
     }
 
-    // Nie można usunąć więcej niż istnieje
     int toRemove = std::min((int)spectrum.size(), lNeg);
 
     auto &rng = RandomGenerator::getInstance().get();
@@ -354,7 +314,6 @@ void PositiveErrorIntroducer::introduceErrors(DNAInstance &instance) {
     static const std::string nucleotides = "ACGT";
     std::unordered_set<std::string> uniqueSpectrum(spectrum.begin(), spectrum.end());
     
-    // Helper do generowania losowego k-meru
     auto generateRandomKmer = [&](int length) {
         std::string kmer;
         kmer.reserve(length);
@@ -364,16 +323,14 @@ void PositiveErrorIntroducer::introduceErrors(DNAInstance &instance) {
         return kmer;
     };
     
-    // Dodajemy błędy pozytywne
     int errorsAdded = 0;
-    int maxAttempts = lPoz * 10; // Limit prób, aby uniknąć nieskończonej pętli
+    int maxAttempts = lPoz * 10;
     int attempts = 0;
     
     while(errorsAdded < lPoz && attempts < maxAttempts) {
         attempts++;
         
         if(probablePositive == 0) {
-            // Generowanie losowej sekwencji
             int d = distDelta(rng);
             int length = k;
             if(d > 0) {
@@ -387,16 +344,14 @@ void PositiveErrorIntroducer::introduceErrors(DNAInstance &instance) {
                 errorsAdded++;
             }
         } else {
-            // Modyfikacja istniejącego oligonukleotydu
             std::uniform_int_distribution<int> indexDist(0, spectrum.size() - 1);
             int idx = indexDist(rng);
             std::string original = spectrum[idx];
             
-            // Modyfikujemy kopię zmieniając jeden losowy nukleotyd
             std::string modified = original;
             int posToChange = (distSign(rng) == 0) ? 
-                modified.length() - 1 :  // ostatni
-                modified.length() / 2;   // środkowy
+                (int)modified.length() - 1 : 
+                (int)modified.length() / 2;
                 
             char originalNuc = modified[posToChange];
             char newNuc;
